+++ CHANGELOG.md
@@ -1,3 +1,59 @@
+# [1.27.0](https://github.com/dxheroes/dx-scanner/compare/v1.26.4...v1.27.0) (2019-12-17)
+
+
+### Bug Fixes
+
+* Add comments ([2c33c17](https://github.com/dxheroes/dx-scanner/commit/2c33c17))
+* add linesAdded and linesRemoved to own interface of Commit. Implement it in BitbucketService and rewrite tests. ([0648ce9](https://github.com/dxheroes/dx-scanner/commit/0648ce9))
+* add pagination to getPullrequests() in Bitbucket ([47e57ad](https://github.com/dxheroes/dx-scanner/commit/47e57ad))
+* adding values ([35f27d5](https://github.com/dxheroes/dx-scanner/commit/35f27d5))
+* concating url ([b0eaaf0](https://github.com/dxheroes/dx-scanner/commit/b0eaaf0))
+* remove unnecessary code ([5dcf879](https://github.com/dxheroes/dx-scanner/commit/5dcf879))
+* remove unnecessary Promise.all(); remove request for getting diffStat ([07e88e5](https://github.com/dxheroes/dx-scanner/commit/07e88e5))
+* rename value ([adcec08](https://github.com/dxheroes/dx-scanner/commit/adcec08))
+* tests after refactoring ([a6f465c](https://github.com/dxheroes/dx-scanner/commit/a6f465c))
+* use method to get an email and fix test after that ([f11ac89](https://github.com/dxheroes/dx-scanner/commit/f11ac89))
+* use pagination if it's provided ([fad8e78](https://github.com/dxheroes/dx-scanner/commit/fad8e78))
+* use withDiffStat value separately, not within ListGetterOptions ([281ab83](https://github.com/dxheroes/dx-scanner/commit/281ab83))
+* value name ([eacdd7b](https://github.com/dxheroes/dx-scanner/commit/eacdd7b))
+
+
+### Features
+
+*  allow to get one specific pull request with diffStat in GithubService and BitbucketService ([837e8a0](https://github.com/dxheroes/dx-scanner/commit/837e8a0))
+* add getPullsDiffStat() in CollaborationInspector; allow to get one specific pull request with diffStat ([81cb265](https://github.com/dxheroes/dx-scanner/commit/81cb265))
+* add Lines interface ([d1232ed](https://github.com/dxheroes/dx-scanner/commit/d1232ed))
+* get pullrequests with diffStat if `withDiffStat` value is true ([dcba226](https://github.com/dxheroes/dx-scanner/commit/dcba226))
+* implement getPullsDiffStat() in BitbucketService and add test for that ([3c78fe3](https://github.com/dxheroes/dx-scanner/commit/3c78fe3))
+
+## [1.26.4](https://github.com/dxheroes/dx-scanner/compare/v1.26.3...v1.26.4) (2019-12-16)
+
+
+### Bug Fixes
+
+* **deps:** update dependency semver to v7 ([76de07d](https://github.com/dxheroes/dx-scanner/commit/76de07d))
+
+## [1.26.3](https://github.com/dxheroes/dx-scanner/compare/v1.26.2...v1.26.3) (2019-12-16)
+
+
+### Bug Fixes
+
+* **deps:** update dependency update-notifier to v4 ([373ce00](https://github.com/dxheroes/dx-scanner/commit/373ce00))
+
+## [1.26.2](https://github.com/dxheroes/dx-scanner/compare/v1.26.1...v1.26.2) (2019-12-16)
+
+
+### Bug Fixes
+
+* **deps:** update all non-major dependencies ([96f59c9](https://github.com/dxheroes/dx-scanner/commit/96f59c9))
+
+## [1.26.1](https://github.com/dxheroes/dx-scanner/compare/v1.26.0...v1.26.1) (2019-12-16)
+
+
+### Bug Fixes
+
+* **deps:** pin dependencies ([57d767a](https://github.com/dxheroes/dx-scanner/commit/57d767a))
+
 # [1.26.0](https://github.com/dxheroes/dx-scanner/compare/v1.25.0...v1.26.0) (2019-12-11)